APPLICATION/app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
opencv-python==4.11.0.86
numpy==2.1.3
python-dotenv==1.1.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
Jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.4
httpx==0.24.1
opentelemetry-instrumentation
opentelemetry-exporter-otlp
ultralytics==8.3.137
python-docx==1.1.2